+++ gui/ui/menu.c
@@ -177,6 +177,6 @@ void uiMenuHide( int mx,int my,int w )
         guiApp.menuItems[i].x+guiApp.menuItems[i].width,
         guiApp.menuItems[i].y+guiApp.menuItems[i].height ) )
    {
-    uiEvent( guiApp.menuItems[i].message,(float)w );
+    uiEvent( guiApp.menuItems[i].message,w );
    }
 }
